Introduce CASH: Camera Augmented Sensing Helper
This is CASH! a Camera Augmented multi-sensor Sensing Helper.
It is composed of a server component and a library that
needs to be included in any HAL (...camera...) that needs
to use this program.

Its aim is to implement, in an opensource way, Laser (ToF)
Auto-Focus and AutoExposure through the RGBC-IR (or
basically any other) sensor.

Currently, the server is dynamically configured with an
XML file and uses libpolyreg to fit the AF data through
the provided polynomial regression data fit algorithm by
using a pre-defined table (in the XML) as starting points.

The current release implements only ToF driven AF.