CASH: Introduce RGBC-IR sensor support
Add a new module that implements basic functionality to control the power state of the sensor and read its values through an input device.
The primary goal is to read the value of the clear channel to get an estimation of the amount of lighting in the scene visible to the back camera.
If the clear value is in the range where camera fails to ramp up brightness, then an ISO value is calcuated from the clear value according to the configuration provided.
ISO value ranges are mapped to fixed exposure times, so CASH is able to inform the camera HAL of both ISO and exposure time to brighten up the scene to an acceptable level.
Due to this, it is also needed to adjust common CASH code to not only return integers through the socket, but structs that contain multiple fields to accomodate for the tuple of ISO and exposure time.

Change-Id: I565ee6388784d61a56b2584be0f338416c890770