CASH: Implement sensors calibration from MiscTA parameters
This commit implements reading the sensors calibration from
their specific MiscTA units and stores them to a file in
the /data/misc/cashsvr/ directory, to avoid opening the TA
at every boot, which can be painfully slow.

This currently reads the parameters for both RGBCIR and
ToF sensors (currently, TCS3490 and STM VL53L0X), but will
only use the parameters for the ToF sensor - for now.

The parameters are sent to the ToF sensor currently via
sysfs are:
- Number of SPADs (type is fixed to not-aperture)
- Ranging offset calibration, expressed in micrometers

These parameters are required because, obviously, the
sensor sits behind a glass window.

While at it, apply some simplification to the chown code.